requirements.txt
pyyaml
Pillow
apscheduler
tk